refactor: Decompose prediction-handler into modular processors and fix resource leaks
